refactor: move URL path builders to a separate module to avoid circular dependencies
